Factor Board class into Game and GameBoard
- Single responsibility principle
- GameBoard takes care of visual representation and layering of tiles
- Game implements the game logic
